Refactor and combine (mostly) into a single package.
